feat: reorganize test and source files by moving tree tests to data_structure
